Trait iced_graphics::backend::Backend

pub trait Backend {
    // Provided method
    fn trim_measurements(&mut self) { ... }
}

The graphics backend of a Renderer.

Provided Methods

fn trim_measurements(&mut self)

Trims the measurements cache.

This method is currently necessary to properly trim the text cache in iced_wgpu and iced_glow because of limitations in the text rendering pipeline. It will be removed in the future.