iced_graphics
Module compositor

A compositor is responsible for initializing a renderer and managing window surfaces.

Structs

Information

Contains information about the graphics (e.g. graphics adapter, graphics backend).

Enums

SurfaceError

Result of an unsuccessful call to Compositor::present.

Traits

Compositor

A graphics compositor that can draw to windows.

Default

Defines the default compositor of a renderer.

Window

A window that can be used in a Compositor.