iced_graphics::compositor

Trait Window

pub trait Window:
    HasWindowHandle
    + HasDisplayHandle
    + MaybeSend
    + MaybeSync
    + 'static { }

A window that can be used in a Compositor.

This is just a convenient super trait of the raw-window-handle traits.