Module iced_graphics::font

Find system fonts or use the built-in ones.

Structs

Source

A font source that can find and load system fonts.

Enums

Familyfont-source

A possible value for the font-family CSS property.

LoadErrorfont-source

Reasons why a source might fail to look up a font or fonts.

Constants

ARROW_DOWN_ICON

The char representing a ▼ icon in the built-in ICONS font.

CHECKMARK_ICONfont-icons

The char representing a ✔ icon in the built-in ICONS font.

FALLBACKfont-fallback

A built-in fallback font, for convenience.

ICONSfont-icons

A built-in icon font, for convenience.