iced_graphics
Module geometry

Build and draw geometry.

Re-exports

pub use fill::Fill;

pub use frame::Frame;

pub use path::Path;

pub use stroke::LineCap;

pub use stroke::LineDash;

pub use stroke::LineJoin;

pub use stroke::Stroke;

pub use crate::gradient;

pub use crate::gradient::Gradient;

Modules

fill

Fill Geometry with a certain style.

frame

Draw and generate geometry.

path

Build different kinds of 2D shapes.

stroke

Create lines from a Path and assigns them various attributes/styles.

Structs

Cache

A simple cache that stores generated geometry to avoid recomputation.

Image

A raster image that can be drawn.

Svg

A raster image that can be drawn.

Text

A bunch of text that can be drawn to a canvas

Enums

Style

The coloring style of some drawing.

Traits

Renderer

A renderer capable of drawing some Self::Geometry.