Module iced_graphics::gradient

For creating a Gradient.

Re-exports

pub use linear::Linear;

Modules

linear

Linear gradient builder & definition.

Structs

ColorStop

A point along the gradient vector where the specified color is unmixed.

Enums

Gradient

A fill which transitions colors progressively along a direction, either linearly, radially (TBD), or conically (TBD).

Location

The location of a relatively-positioned gradient.

Position

The position of the gradient within its bounds.