Enum iced_graphics::primitive::Primitive

#[non_exhaustive]
pub enum Primitive {
    Text {
        content: String,
        bounds: Rectangle,
        color: Color,
        size: f32,
        line_height: LineHeight,
        font: Font,
        horizontal_alignment: Horizontal,
        vertical_alignment: Vertical,
        shaping: Shaping,
    },
    Quad {
        bounds: Rectangle,
        background: Background,
        border_radius: [f32; 4],
        border_width: f32,
        border_color: Color,
    },
    Image {
        handle: Handle,
        bounds: Rectangle,
    },
    Svg {
        handle: Handle,
        color: Option<Color>,
        bounds: Rectangle,
    },
    SolidMesh {
        buffers: Mesh2D<ColoredVertex2D>,
        size: Size,
    },
    GradientMesh {
        buffers: Mesh2D<GradientVertex2D>,
        size: Size,
    },
    Fill {
        path: Path,
        paint: Paint<'static>,
        rule: FillRule,
        transform: Transform,
    },
    Stroke {
        path: Path,
        paint: Paint<'static>,
        stroke: Stroke,
        transform: Transform,
    },
    Group {
        primitives: Vec<Primitive>,
    },
    Clip {
        bounds: Rectangle,
        content: Box<Primitive>,
    },
    Translate {
        translation: Vector,
        content: Box<Primitive>,
    },
    Cache {
        content: Arc<Primitive>,
    },
}

A rendering primitive.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Text

Fields

content: String

The contents of the text

bounds: Rectangle

The bounds of the text

color: Color

The color of the text

size: f32

The size of the text in logical pixels

line_height: LineHeight

The line height of the text

font: Font

The font of the text

horizontal_alignment: Horizontal

The horizontal alignment of the text

vertical_alignment: Vertical

The vertical alignment of the text

shaping: Shaping

The shaping strategy of the text.

A text primitive

Quad

Fields

bounds: Rectangle

The bounds of the quad

background: Background

The background of the quad

border_radius: [f32; 4]

The border radii of the quad

border_width: f32

The border width of the quad

border_color: Color

The border color of the quad

A quad primitive

Image

Fields

handle: Handle

The handle of the image

bounds: Rectangle

The bounds of the image

An image primitive

Svg

Fields

handle: Handle

The path of the SVG file

color: Option<Color>

The [Color] filter

bounds: Rectangle

The bounds of the viewport

An SVG primitive

SolidMesh

Fields

buffers: Mesh2D<ColoredVertex2D>

The vertices and indices of the mesh.

size: Size

The size of the drawable region of the mesh.

Any geometry that falls out of this region will be clipped.

A low-level primitive to render a mesh of triangles with a solid color.

It can be used to render many kinds of geometry freely.

GradientMesh

Fields

buffers: Mesh2D<GradientVertex2D>

The vertices and indices of the mesh.

size: Size

The size of the drawable region of the mesh.

Any geometry that falls out of this region will be clipped.

A low-level primitive to render a mesh of triangles with a gradient.

It can be used to render many kinds of geometry freely.

Fill

Fields

path: Path

The path to fill.

paint: Paint<'static>

The paint to use.

rule: FillRule

The fill rule to follow.

transform: Transform

The transform to apply to the path.

Available on crate feature tiny-skia only.

A [tiny_skia] path filled with some paint.

Stroke

Fields

path: Path

The path to stroke.

paint: Paint<'static>

The paint to use.

stroke: Stroke

The stroke settings.

transform: Transform

The transform to apply to the path.

Available on crate feature tiny-skia only.

A [tiny_skia] path stroked with some paint.

Group

Fields

primitives: Vec<Primitive>

The primitives of the group

A group of primitives

Clip

Fields

bounds: Rectangle

The bounds of the clip

content: Box<Primitive>

The content of the clip

A clip primitive

Translate

Fields

translation: Vector

The translation vector

content: Box<Primitive>

The primitive to translate

A primitive that applies a translation

Cache

Fields

content: Arc<Primitive>

The cached primitive

A cached primitive.

This can be useful if you are implementing a widget where primitive generation is expensive.

Implementations

impl Primitive

pub fn group(primitives: Vec<Self>) -> Self

Creates a Primitive::Group.

pub fn clip(self, bounds: Rectangle) -> Self

Creates a Primitive::Clip.

pub fn translate(self, translation: Vector) -> Self

Creates a Primitive::Translate.

pub fn bounds(&self) -> Rectangle

Returns the bounds of the Primitive.