Module iced_graphics::widget::canvas

Available on crate feature canvas only.

Draw 2D graphics for your users.

A Canvas widget can be used to draw different kinds of 2D shapes in a Frame. It can be used for animation, data visualization, game graphics, and more!

Re-exports

pub use crate::gradient;

pub use crate::gradient::Gradient;

pub use event::Event;

pub use fill::Fill;

pub use fill::FillRule;

pub use path::Path;

pub use stroke::LineCap;

pub use stroke::LineDash;

pub use stroke::LineJoin;

pub use stroke::Stroke;

Modules

event

Handle events of a canvas.

fill

Fill crate::widget::canvas::Geometry with a certain style.

path

Build different kinds of 2D shapes.

stroke

Create lines from a crate::widget::canvas::Path and assigns them various attributes/styles.

Structs

Cache

A simple cache that stores generated Geometry to avoid recomputation.

Canvas

A widget capable of drawing 2D graphics.

Frame

The frame of a Canvas.

Geometry

A bunch of shapes that can be drawn.

Text

A bunch of text that can be drawn to a canvas

Enums

Cursor

The mouse cursor state.

Traits

Program

The state and logic of a Canvas.