Module iced_graphics::widget::canvas::fill

Available on crate feature canvas only.

Fill crate::widget::canvas::Geometry with a certain style.

Re-exports

pub use crate::triangle::Style;

Structs

Fill

The style used to fill geometry.

Enums

FillRule

The fill rule defines how to determine what is inside and what is outside of a shape.