Trait iced_graphics::widget::pure::canvas::_

pub trait _<Message> {
    fn draw(&self, bounds: Rectangle, cursor: Cursor) -> Vec<Geometry>;

    fn update(
        &mut self, 
        _event: Event, 
        _bounds: Rectangle, 
        _cursor: Cursor
    ) -> (Status, Option<Message>) { ... }
    fn mouse_interaction(
        &self, 
        _bounds: Rectangle, 
        _cursor: Cursor
    ) -> Interaction { ... }
}

The state and logic of a Canvas.

A Program can mutate internal state and produce messages for an application.

Required Methods

fn draw(&self, bounds: Rectangle, cursor: Cursor) -> Vec<Geometry>

Draws the state of the Program, producing a bunch of Geometry.

Geometry can be easily generated with a Frame or stored in a Cache.

Provided Methods

fn update(
&mut self,
_event: Event,
_bounds: Rectangle,
_cursor: Cursor
) -> (Status, Option<Message>)

Updates the state of the Program.

When a Program is used in a Canvas, the runtime will call this method for each Event.

This method can optionally return a Message to notify an application of any meaningful interactions.

By default, this method does and returns nothing.

fn mouse_interaction(&self, _bounds: Rectangle, _cursor: Cursor) -> Interaction

Returns the current mouse interaction of the Program.

The interaction returned will be in effect even if the cursor position is out of bounds of the program’s Canvas.