Trait iced_graphics::compositor::Compositor

pub trait Compositor: Sized {
    type Settings: Default;
    type Renderer: Renderer;
    type Surface;

    // Required methods
    fn new<W: HasRawWindowHandle + HasRawDisplayHandle>(
        settings: Self::Settings,
        compatible_window: Option<&W>
    ) -> Result<(Self, Self::Renderer), Error>;
    fn create_surface<W: HasRawWindowHandle + HasRawDisplayHandle>(
        &mut self,
        window: &W,
        width: u32,
        height: u32
    ) -> Self::Surface;
    fn configure_surface(
        &mut self,
        surface: &mut Self::Surface,
        width: u32,
        height: u32
    );
    fn fetch_information(&self) -> Information;
    fn present<T: AsRef<str>>(
        &mut self,
        renderer: &mut Self::Renderer,
        surface: &mut Self::Surface,
        viewport: &Viewport,
        background_color: Color,
        overlay: &[T]
    ) -> Result<(), SurfaceError>;
    fn screenshot<T: AsRef<str>>(
        &mut self,
        renderer: &mut Self::Renderer,
        surface: &mut Self::Surface,
        viewport: &Viewport,
        background_color: Color,
        overlay: &[T]
    ) -> Vec<u8>;
}

A graphics compositor that can draw to windows.

Required Associated Types

type Settings: Default

The settings of the backend.

type Renderer: Renderer

The iced renderer of the backend.

type Surface

The surface of the backend.

Required Methods

fn new<W: HasRawWindowHandle + HasRawDisplayHandle>( settings: Self::Settings, compatible_window: Option<&W> ) -> Result<(Self, Self::Renderer), Error>

Creates a new Compositor.

fn create_surface<W: HasRawWindowHandle + HasRawDisplayHandle>( &mut self, window: &W, width: u32, height: u32 ) -> Self::Surface

Crates a new Surface for the given window.

fn configure_surface( &mut self, surface: &mut Self::Surface, width: u32, height: u32 )

Configures a new Surface with the given dimensions.

fn fetch_information(&self) -> Information

Returns Information used by this Compositor.

fn present<T: AsRef<str>>( &mut self, renderer: &mut Self::Renderer, surface: &mut Self::Surface, viewport: &Viewport, background_color: Color, overlay: &[T] ) -> Result<(), SurfaceError>

Presents the Renderer primitives to the next frame of the given Surface.

fn screenshot<T: AsRef<str>>( &mut self, renderer: &mut Self::Renderer, surface: &mut Self::Surface, viewport: &Viewport, background_color: Color, overlay: &[T] ) -> Vec<u8>

Screenshots the current Renderer primitives to an offscreen texture, and returns the bytes of the texture ordered as RGBA in the sRGB color space.