Struct iced_graphics::renderer::Renderer

pub struct Renderer<B: Backend> { /* fields omitted */ }

A backend-agnostic renderer that supports all the built-in widgets.

Implementations

impl<B: Backend> Renderer<B>

pub fn new(backend: B) -> Self

Creates a new Renderer from the given Backend.

pub fn backend(&self) -> &B

Returns the Backend of the Renderer.

pub fn draw_primitive(&mut self, primitive: Primitive)

Enqueues the given Primitive in the Renderer for drawing.

pub fn with_primitives(&mut self, f: impl FnOnce(&mut B, &[Primitive]))

Runs the given closure with the Backend and the recorded primitives of the Renderer.