Struct iced_graphics::renderer::Renderer

pub struct Renderer<B: Backend> { /* private fields */ }

A backend-agnostic renderer that supports all the built-in widgets.

Implementations

impl<B: Backend> Renderer<B>

pub fn new(backend: B, default_font: Font, default_text_size: Pixels) -> Self

Creates a new Renderer from the given Backend.

pub fn backend(&self) -> &B

Returns a reference to the Backend of the Renderer.

pub fn draw_primitive(&mut self, primitive: Primitive<B::Primitive>)

Enqueues the given Primitive in the Renderer for drawing.

pub fn with_primitives<O>( &mut self, f: impl FnOnce(&mut B, &[Primitive<B::Primitive>]) -> O ) -> O

Runs the given closure with the Backend and the recorded primitives of the Renderer.