Struct iced_graphics::renderer::Renderer

pub struct Renderer<B: Backend, Theme> { /* private fields */ }

A backend-agnostic renderer that supports all the built-in widgets.

Implementations

impl<B: Backend, T> Renderer<B, T>

pub fn new(backend: B) -> Self

Creates a new Renderer from the given Backend.

pub fn backend(&self) -> &B

Returns a reference to the Backend of the Renderer.

pub fn draw_primitive(&mut self, primitive: Primitive<B::Primitive>)

Enqueues the given Primitive in the Renderer for drawing.

pub fn with_primitives<O>( &mut self, f: impl FnOnce(&mut B, &[Primitive<B::Primitive>]) -> O ) -> O

Runs the given closure with the Backend and the recorded primitives of the Renderer.

pub fn start_layer(&mut self) -> Vec<Primitive<B::Primitive>>

Starts recording a new layer.

pub fn end_layer( &mut self, primitives: Vec<Primitive<B::Primitive>>, bounds: Rectangle )

Ends the recording of a layer.

pub fn start_translation(&mut self) -> Vec<Primitive<B::Primitive>>

Starts recording a translation.

pub fn end_translation( &mut self, primitives: Vec<Primitive<B::Primitive>>, translation: Vector )

Ends the recording of a translation.