Module iced_graphics::widget::canvas

Available on crate feature canvas only.

Draw 2D graphics for your users.

A Canvas widget can be used to draw different kinds of 2D shapes in a Frame. It can be used for animation, data visualization, game graphics, and more!

Re-exports

pub use event::Event;

pub use path::Path;

Modules

event

Handle events of a canvas.

path

Build different kinds of 2D shapes.

Structs

Cache

A simple cache that stores generated Geometry to avoid recomputation.

Canvas

A widget capable of drawing 2D graphics.

Fill

The style used to fill geometry.

Frame

The frame of a Canvas.

Geometry

A bunch of shapes that can be drawn.

LineDash

The dash pattern used when stroking the line.

Stroke

The style of a stroke.

Text

A bunch of text that can be drawn to a canvas

Enums

Cursor

The mouse cursor state.

FillRule

The fill rule defines how to determine what is inside and what is outside of a shape.

LineCap

The shape used at the end of open subpaths when they are stroked.

LineJoin

The shape used at the corners of paths or basic shapes when they are stroked.

Traits

Program

The state and logic of a Canvas.